assert right https header
